Rename useDelay -> useDelayOnce
