+++ packages/react-jsx-highcharts/src/components/UseDelayOnce/index.js
